+++ index.js
@@ -1,5 +1,0 @@
-var path = require('path');
-
-module.exports = {
-  includePath: path.join(__dirname, 'src'),
-};